Add `clamp_cast<T>(U)`
* Static casts are mostly UB when the input type value is out of the output
  type range. For IO functions, this is sometimes annoying, not because users
  need to make sure their values are clamp and will fit in the destination
  type, but because they have to figure out the min and max value of the
  data type they use for (de)serialization. This is not particularly difficult
  but it is very impractical. As such, `clamp_cast` was added to safely cast
  any "data type" (defined by the type traits `is_data`) to any other "data
  type". In most cases, this is as or almost as efficient as a `static_cast`.

* Basic type traits have been updated to include ALL integral types in the
  `is_int` traits. `is_sint` and `is_uint` targets signed and unsigned
  integers, respectively.

* The "long" aliases of the Int static template vectors uses fixed-size
  integers now, i.e. `int64_t` and `uint64_t` (before they were using `long`
  and `long long`). In this case, the fixed-size integers seem to be better
  since it is supposed to be the "preferred" 8-bytes integer type of the
  platform. As mentioned in `Types.h`, the code base is organized in such a
  way that it should not really matter and either `long` or `long long` can be
  used (on platform with `sizeof(long)==8`) for 8-bytes integers.

Tags: [feature][update]
Compile: no
Tests: no

+++ src/noa/common/types/Int2.h
@@ -125,8 +125,8 @@ namespace noa {
 
     using int2_t = Int2<int>;
     using uint2_t = Int2<uint>;
-    using long2_t = Int2<long long>;
-    using ulong2_t = Int2<unsigned long long>;
+    using long2_t = Int2<int64_t>;
+    using ulong2_t = Int2<uint64_t>;
 
     template<typename T>
     NOA_IH constexpr std::array<T, 2> toArray(const Int2<T>& v) noexcept {
