+++ src/noa/common/types/Int4.h
@@ -142,8 +142,8 @@ namespace noa {
 
     using int4_t = Int4<int>;
     using uint4_t = Int4<uint>;
-    using long4_t = Int4<long long>;
-    using ulong4_t = Int4<unsigned long long>;
+    using long4_t = Int4<int64_t>;
+    using ulong4_t = Int4<uint64_t>;
 
     template<typename T>
     NOA_IH constexpr std::array<T, 4> toArray(const Int4<T>& v) noexcept {
